+++ tensorflow_similarity/losses.py
@@ -26,17 +26,17 @@ References
 
 import tensorflow as tf
 from .utils import is_tensor_or_variable
-from .distances import metric_name_canonializer, pairwise_cosine
+from .distances import Distance, distance_canonicalizer
 from .algebra import masked_maximum, masked_minimum, build_masks
-from .types import FloatTensorLike, Distance
-from typing import List, Union
+from .types import FloatTensor
+from typing import Callable, List, Union
 
 
 @tf.keras.utils.register_keras_serializable(package="Similarity")
 @tf.function
 def triplet_loss(labels: List[int],
-                 embeddings: FloatTensorLike,
-                 distance: Union[Distance, str] = 'cosine',
+                 embeddings: FloatTensor,
+                 distance: Callable,
                  positive_mining_strategy: str = 'hard',
                  negative_mining_strategy: str = 'semi-hard',
                  soft_margin: bool = False,
@@ -70,11 +70,7 @@ def triplet_loss(labels: List[int],
     batch_size = tf.size(labels)
 
     # [distances]
-    if distance == 'cosine':
-        pairwise_distances = pairwise_cosine(embeddings)
-    else:
-        # user supplied distance function
-        pairwise_distances = distance(embeddings)
+    pairwise_distances = distance(embeddings)
 
     # [masks]
     positive_mask, negative_mask = build_masks(labels, batch_size)
@@ -230,7 +226,7 @@ class TripletLoss(LossFunctionWrapper):
         """
 
         # distance canonicalization
-        distance = metric_name_canonializer(distance)
+        distance = distance_canonicalizer(distance)
         self.distance = distance
         # sanity checks
 
@@ -243,7 +239,7 @@ class TripletLoss(LossFunctionWrapper):
         # Ensure users knows its one or the other
         if margin != 1.0 and soft_margin:
             raise ValueError('Margin value is not used when soft_margin is\
-                set to True'                            )
+                              set to True')
 
         super().__init__(triplet_loss,
                          name=name,