+++ tensorflow_similarity/matchers/nmslib_matcher.py
@@ -2,15 +2,18 @@ import nmslib
 from pathlib import Path
 from typing import List, Tuple
 
+from tensorflow_similarity.distances import distance_canonicalizer
 from .matcher import Matcher
-from tensorflow_similarity.types import FloatTensorLike
+from tensorflow_similarity.types import FloatTensor
 
 
 class NMSLibMatcher(Matcher):
 
     def __init__(self, distance='cosine', algorithm='nmslib_hnsw', verbose=1):
 
-        if distance == 'cosine':
+        distance = distance_canonicalizer(distance)
+
+        if distance.name == 'cosine':
             space = 'cosinesimil'
         else:
             raise ValueError('Unsupported metric space')
@@ -26,14 +29,14 @@ class NMSLibMatcher(Matcher):
         self._matcher.createIndex(print_progress=show_prog)
 
     def add(self,
-            embedding: FloatTensorLike,
+            embedding: FloatTensor,
             idx: int,
             build: bool = True,
             verbose: int = 1):
         """Add an embedding to the index
 
         Args:
-            embedding (FloatTensorLike): Record embedding as computed
+            embedding (FloatTensor): Record embedding as computed
             by the model.
 
             idx (int): Embedding id in the index table. Used to lookup
@@ -50,14 +53,14 @@ class NMSLibMatcher(Matcher):
             self._build(verbose=verbose)
 
     def batch_add(self,
-                  embeddings: FloatTensorLike,
+                  embeddings: FloatTensor,
                   idxs: List[int],
                   build: bool = True,
                   verbose: int = 1):
         """Add an embedding to the index
 
         Args:
-            embeddings (FloatTensorLike): List of embeddings to add to the
+            embeddings (FloatTensor): List of embeddings to add to the
             index.
 
             idxs (int): Embedding id in the index table. Used to lookup
@@ -80,12 +83,12 @@ class NMSLibMatcher(Matcher):
                 print('|-Optimizing NN matcher index.')
 
     def lookup(self,
-               embedding: FloatTensorLike,
+               embedding: FloatTensor,
                k: int = 5) -> Tuple[List[int], List[float]]:
         """Find the embedding K nearest neighboors
 
         Args:
-            embedding (FloatTensorLike): Target embedding as predicted by
+            embedding (FloatTensor): Target embedding as predicted by
             the model.
             k (int, optional): Number of nearest neighboors to lookup.
             Defaults to 5.
@@ -96,11 +99,11 @@ class NMSLibMatcher(Matcher):
         return idxs, distances
 
     def batch_lookup(self,
-                     embeddings: FloatTensorLike,
+                     embeddings: FloatTensor,
                      k: int = 5) -> Tuple[List[List[int]], List[List[float]]]:
         """Find embeddings K nearest neighboors
         Args:
-            embedding (FloatTensorLike): Target embedding as predicted by
+            embedding (FloatTensor): Target embedding as predicted by
             the model.
             k (int, optional): Number of nearest neighboors to lookup.
             Defaults to 5.