+++ tests/test_distances.py
@@ -1,9 +1,8 @@
 import numpy as np
 import tensorflow as tf
 
-from tensorflow_similarity.distances import cosine
-from tensorflow_similarity.distances import pairwise_cosine
-from tensorflow_similarity.distances import pairwise_euclidean
+from tensorflow_similarity.distances import CosineDistance
+from tensorflow_similarity.distances import EuclidianDistance
 
 
 def angular_distance_np(feature):
@@ -26,42 +25,42 @@ def angular_distance_np(feature):
 def test_cosine_same():
     # pairwise
     a = tf.convert_to_tensor([[1.0, 1.0, 1.0], [1.0, 1.0, 1.0]])
-    vals = pairwise_cosine(a)
+    d = CosineDistance()
+    vals = d(a)
     assert tf.round(tf.reduce_sum(vals)) == 0
 
 
 def test_cosine_opposite():
     a = tf.convert_to_tensor([[0.0, 1.0], [1.0, 0.0]])
-    vals = pairwise_cosine(a)
+    d = CosineDistance()
+    vals = d(a)
     assert tf.round(tf.reduce_sum(vals)) == 2
 
 
 def test_cosine_vals():
     a = tf.constant([[0.1, 0.3, 0.2], [0.0, 0.1, 0.5]])
-    vals = pairwise_cosine(a)
+    d = CosineDistance()
+    vals = d(a)
     assert vals[0][0] == 0
     assert vals[0][1] == 0.31861484
 
 
-def test_direct_cosine():
-    a = tf.constant([[0.1, 0.3, 0.2]])
-    b = tf.constant([[0.0, 0.1, 0.5]])
-    assert cosine(a, b) == 0.31861484
-
-
 def test_euclidean():
     a = tf.convert_to_tensor([[0.0, 3.0], [4.0, 0.0]])
-    vals = pairwise_euclidean(a)
+    d = EuclidianDistance()
+    vals = d(a)
     assert tf.reduce_all(tf.math.equal(vals, tf.constant([[1e-8, 5.0],[5.0, 1e-8]])))
 
 
 def test_euclidean_same():
     a = tf.convert_to_tensor([[1.0, 1.0], [1.0, 1.0]])
-    vals = pairwise_euclidean(a)
+    d = EuclidianDistance()
+    vals = d(a)
     assert tf.round(tf.reduce_sum(vals)) == 0
 
 
 def test_euclidean_opposite():
     a = tf.convert_to_tensor([[0.0, 1.0], [0.0, -1.0]])
-    vals = pairwise_euclidean(a)
+    d = EuclidianDistance()
+    vals = d(a)
     assert tf.reduce_all(tf.math.equal(vals, tf.constant([[1e-8, 2.0],[2.0, 1e-8]])))